pytest_appium/requirements.txt
selenium
pytest
Appium-Python-Client
